small refacto
should allow to have a clean search state to search on different pages
